add_hw_RecyclerView

+++ app/src/main/java/Post.kt
@@ -5,7 +5,7 @@ data class Post (
     val author: String,
     val content: String,
     val published: String,
-    var likes: Int,
-    var reposts: Int,
+    var likes: Long,
+    var reposts: Long,
     val likedByMe: Boolean
 )
